POCs/RealTime-Predictive-System/requirements.txt
pyspark
mysql-connector-python
psycopg2-binary
kafka-python
apache-airflow
pandas
scikit-learn